+++ src/constants/storageKeys.ts
@@ -1,0 +1,3 @@
+export const AUTH_TOKEN_KEY = "auth_token";
+export const USER_KEY = "user";
+export const OTP_VERIFIED_KEY = "otp_verified";