+++ src/shared/component/ProtectedRoutes.tsx
@@ -3,12 +3,16 @@ import { Navigate } from "react-router-dom";
 import { useAuth } from "../hooks/useAuth";
 
 function ProtectedRoutes({ children }: { children: React.ReactNode }) {
-  const { user, token, ready } = useAuth();
+  const { user, token, otpVerified, ready } = useAuth();
 
   if (!ready) return null;
 
   if (!user || !token) {
     return <Navigate to="/sign-in" replace />;
+  }
+
+  if (!otpVerified) {
+    return <Navigate to="/verify-otp" replace />;
   }
 
   return <>{children}</>;