+++ CHANGELOG.md
@@ -32,7 +32,7 @@
 
 ## [2.1.0] 2022-02-14
 
-- TypeScript version of the product is available: [Tajhiz](https://www.creative-tim.com/product/material-dashboard-2-pro-react-ts)
+- TypeScript version of the product is available: [Meena-health](https://www.creative-tim.com/product/material-dashboard-2-pro-react-ts)
 
 ### Bug fixing